낙관적 lock도입

+++ src/main/java/com/twogather/twogatherwebbackend/dto/store/TopStoreResponse.java
@@ -13,9 +13,9 @@ public class TopStoreResponse{
     private String address;
     private Double avgScore;
     private String storeImageUrl;
-    private Integer likeCount;
+    private Long likeCount;
     public TopStoreResponse(Long storeId, String name, Double avgScore,
-                            String address, String storeImageUrl, Integer likeCount){
+                            String address, String storeImageUrl, long likeCount){
         this.storeId = storeId;
         this.storeName = name;
         this.address = address;
